Included metabox script and began updating phpdoc throughout.
Kudos to
https://github.com/jaredatch/Custom-Metaboxes-and-Fields-for-WordPress
for a great meta box script, happy to include it and rely on it. Still
toying around, but excited for it. Porting over my (really) custom
metabox setup in favor of this one. Also inspired by the solid phpdoc
throughout, so going to be updating some of mine. Likely improving code
in the process :)